Still geting log probability calculation issues

+++ Scripts/Custom_Functions/dhyper_nimble.R
@@ -1,6 +1,11 @@
 # Define Hypergeometric Sampling Function from R
 dhyper_nimble <- nimbleRcall(
-  function(x = double(0), m = double(0), n = double(0), k = double(0), log = integer(0, default = 0)){},
+  # Have to use R notation, but for model
+  # x = manually validated calls found to be true (k)
+  # m = number of true detections (K)
+  # n = (is n in R) number of false positive detections (Q)
+  # k = number of calls validated (n)
+  function(x = double(0), m = double(0), n = double(0), k = double(0), log = double(0, default = 0)){},
   Rfun = 'dhyper',
   returnType = double()
 )
